+++ src/components/Fav/SongList/SongListTableActions.tsx
@@ -9,7 +9,7 @@ import KeyboardArrowLeft from '@mui/icons-material/KeyboardArrowLeft';
 import KeyboardArrowRight from '@mui/icons-material/KeyboardArrowRight';
 import LastPageIcon from '@mui/icons-material/LastPage';
 
-import { skinPreset } from '../../styles/skin';
+import { skinPreset } from '../../../styles/skin';
 
 const { colorTheme } = skinPreset;
